Create temporary files for keystore and group policy
