+++ applications/workers/workers-smoketest/src/smokeTest/kotlin/net/corda/applications/workers/smoketest/virtualnode/helpers/ClusterBuilder.kt
@@ -70,7 +70,7 @@ class ClusterBuilder {
         uploadCpiResource("/api/v1/cpi/", resourceName, groupId, staticMemberNames, cpiName)
 
     fun updateVirtualNodeState(holdingIdHash: String, newState: String) =
-        put("/api/v1/maintenance/virtualnode/$holdingIdHash/state/$newState", "")
+        put("/api/v1/virtualnode/$holdingIdHash/state/$newState", "")
 
     /** Assumes the resource is a CPB and converts it to CPI by adding a group policy file */
     fun forceCpiUpload(resourceName: String, groupId: String, staticMemberNames: List<String>, cpiName: String) =
