+++ libs/virtual-node/virtual-node-endpoints/src/main/kotlin/net/corda/libs/virtualnode/endpoints/v1/VirtualNodeRPCOps.kt
@@ -3,9 +3,11 @@ package net.corda.libs.virtualnode.endpoints.v1
 import net.corda.httprpc.RpcOps
 import net.corda.httprpc.annotations.HttpRpcGET
 import net.corda.httprpc.annotations.HttpRpcPOST
+import net.corda.httprpc.annotations.HttpRpcPUT
 import net.corda.httprpc.annotations.HttpRpcRequestBodyParameter
-import net.corda.httprpc.annotations.HttpRpcResource
 import net.corda.httprpc.annotations.HttpRpcPathParameter
+import net.corda.httprpc.annotations.HttpRpcResource
+import net.corda.libs.virtualnode.endpoints.v1.types.ChangeVirtualNodeStateResponse
 import net.corda.libs.virtualnode.endpoints.v1.types.HoldingIdentity
 import net.corda.libs.virtualnode.endpoints.v1.types.VirtualNodeRequest
 import net.corda.libs.virtualnode.endpoints.v1.types.VirtualNodes
@@ -46,6 +48,26 @@ interface VirtualNodeRPCOps : RpcOps {
         responseDescription = "List of virtual node details."
     )
     fun getAllVirtualNodes(): VirtualNodes
+
+    /**
+     * Updates a virtual nodes state.
+     *
+     * @throws `VirtualNodeRPCOpsServiceException` If the virtual node update request could not be published.
+     * @throws `HttpApiException` If the request returns an exceptional response.
+     */
+    @HttpRpcPUT(
+        path = "{virtualNodeShortId}/state/{newState}",
+        title = "Update virtual node state",
+        description = "This method updates the state of a new virtual node to one of the pre-defined values.",
+        responseDescription = "Complete information about updated virtual node which will also contain the updated state."
+    )
+    fun updateVirtualNodeState(
+        @HttpRpcPathParameter(description = "Short ID of the virtual node instance to update")
+        virtualNodeShortId: String,
+        @HttpRpcPathParameter(description = "State to transition virtual node instance into. " +
+                "Possible values are: IN_MAINTENANCE and ACTIVE.")
+        newState: String
+    ): ChangeVirtualNodeStateResponse
 
     /**
      * Returns the VirtualNodeInfo for a given [HoldingIdentity].
